third_party/WebKit/LayoutTests/fast/html/imports/import-style-link-block-expected.txt

CONSOLE WARNING: Styling master document from stylesheets defined in HTML Imports is deprecated, and is planned to be removed in M67, around May 2018. Please refer to https://goo.gl/EGXzpw for possible migration paths.
PASS childStyleLink.sheet.title is "Script Blocker"
PASS importLink.import is non-null.
PASS successfullyParsed is true

TEST COMPLETE
